Fix handling of blank addresses
The previous code here would treat a blank address as if it said `*`.
However, this was not consistent across the rest of the mod, so bees
would repeatedly pick up the package just to place it back into the
same bee port. I changed it to look for a bee port with a blank name,
which matches how trains and frogports do it.

Fixes #184